requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-jwt-auth==0.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
